requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.5.0
apscheduler==3.10.4
streamlit==1.31.0
requests==2.31.0
pandas==2.2.0